+++ app/src/main/java/ru/ls/cardriver/domain/model/PointLocation.kt
@@ -2,7 +2,7 @@ package ru.ls.cardriver.domain.model
 
 import android.graphics.Point
 
-data class PointLocation(val x: Float, val y: Float) {
+data class PointLocation(val x: Int, val y: Int) {
 
-	fun toPoint(): Point = Point(x.toInt(), y.toInt())
+	fun toPoint(): Point = Point(x, y)
 }
